fastapi/pyproject.toml
[tool.poetry]
name = "prediction-price-fastapi"
version = "0.1.0"
description = ""
authors = ["jeonghun <p65454@gmail.com>"]
readme = "README.md"
packages = [{include = "fastapi"}]

[tool.poetry.dependencies]
python = "^3.11"
numpy = "^1.26.4"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.29"}
pendulum = "^3.0.0"
pyupbit = "^0.2.33"
uvicorn = {extras = ["standard"], version = "^0.30.1"}
fastapi = "^0.110.2"
mlflow = "1.27.0"
psycopg2-binary = "2.9.9"
boto3 = "1.34.78"
pydantic-settings = "^2.3.1"
asyncpg = "^0.29.0"
orjson = "^3.10.3"
minio = "^7.2.7"
catboost = "^1.2.5"
python-dotenv = "^1.0.1"
redis = "^5.0.6"
starlette-context = "^0.3.6"
dependency-injector = "^4.41.0"
aiohttp = "^3.9.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
